+++ apps/web/src/components/sections/PostsSection.vue
@@ -8,7 +8,7 @@ defineProps<{
 </script>
 
 <template>
-  <section class="sec">
+  <section :id="module.id" class="sec">
     <div class="sec-head">
       <h2>{{ module.data.heading }}</h2>
     </div>